Rework wait(,_conn)_async() functions
We use Future objects instead Event in async waiting functions. This
allows for both the reader and the writer to return ready, as it is
already possible in non-async wait*() functions.

In wait_conn_async(), we need catch timeout error when no future is done
(as documented in asyncio.wait()).