Add unit tests for wait*() functions
We check that when a generator waits for a Wait value, it gets a Ready
value that matches. The socket we wait on is read- and write-ready.

These tests hang on non-Linux platform in CI, perhaps because some
socket operations (e.g. fileno()) are not portable, so we only run them
on Linux.

As is, wait_epoll() fails this test because it assumes that readiness is
either read or write, not both.